move EnrichedEvent to node-commons
