+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/event/common/heartbeat/PipeHeartbeatEvent.java
@@ -21,10 +21,10 @@ package org.apache.iotdb.db.pipe.event.common.heartbeat;
 
 import org.apache.iotdb.commons.consensus.index.ProgressIndex;
 import org.apache.iotdb.commons.consensus.index.impl.MinimumProgressIndex;
+import org.apache.iotdb.commons.pipe.event.EnrichedEvent;
 import org.apache.iotdb.commons.pipe.task.connection.BoundedBlockingPendingQueue;
 import org.apache.iotdb.commons.pipe.task.connection.UnboundedBlockingPendingQueue;
 import org.apache.iotdb.commons.pipe.task.meta.PipeTaskMeta;
-import org.apache.iotdb.db.pipe.event.EnrichedEvent;
 import org.apache.iotdb.db.pipe.extractor.realtime.PipeRealtimeDataRegionExtractor;
 import org.apache.iotdb.db.pipe.extractor.realtime.PipeRealtimeDataRegionHybridExtractor;
 import org.apache.iotdb.db.pipe.metric.PipeHeartbeatEventMetrics;
@@ -208,6 +208,14 @@ public class PipeHeartbeatEvent extends EnrichedEvent {
 
   public void bindExtractor(PipeRealtimeDataRegionExtractor extractor) {
     this.extractor = extractor;
+  }
+
+  /////////////////////////////// For Commit Ordering ///////////////////////////////
+
+  /** HeartbeatEvents do not need to be committed in order. */
+  @Override
+  public boolean needToCommit() {
+    return false;
   }
 
   /////////////////////////////// Object ///////////////////////////////
